fix: prevent NoneType TypeError in pandas arithmetic operations
- safe_add_features(): move df.copy() before the coercion loop (was
  after), add .astype(float) to all pd.to_numeric() calls on OHLCV
  columns. Without .astype(float), a column with mixed Python None /
  float values stays in object dtype after pd.to_numeric(errors='coerce'),
  so pandas falls back to Python-level ops and raises:
    TypeError: unsupported operand type(s) for -: 'float' and 'NoneType'

- compute_regime_and_trend(): apply the same pd.to_numeric().astype(float)
  sanitization to high/low/close before the True Range pd.concat()
  arithmetic, using local sanitized Series instead of raw df columns.

- feature_engineering.add_features(): same .astype(float) fix on the
  per-column coercion loop so None doesn't survive into pandas_ta's
  internal ATR/MACD subtraction path.

+++ feature_engineering.py
@@ -46,7 +46,11 @@ def add_features(df: pd.DataFrame) -> pd.DataFrame:
             for feature in FEATURE_COLS:
                 temp_df[feature] = FEATURE_DEFAULTS.get(feature, 0.0) 
             return temp_df
-        df_copy[col] = pd.to_numeric(df_copy[col], errors='coerce')
+        # FIX: .astype(float) evicts Python None from object-dtype columns;
+        # without it, pd.to_numeric(errors='coerce') can leave None in place when
+        # the column dtype stays 'object', causing:
+        #   TypeError: unsupported operand type(s) for -: 'float' and 'NoneType'
+        df_copy[col] = pd.to_numeric(df_copy[col], errors='coerce').astype(float)
 
     # Calculate 'returns'
     df_copy['returns'] = df_copy['close'].pct_change().fillna(0)
